# server/openapi.yaml
openapi: 3.1.0
info:
  title: HiChat API
  version: "0.1.0"
  description: |-
    REST + WebSocket API for HiChat, a Discord-style chat platform (servers, channels,
    DMs, voice/video via LiveKit, end-to-end-encrypted messaging, a music/soundboard
    bot, and a read-only Bot Gateway).

    ## Response envelope
    Every JSON response — success or error — is wrapped in the same envelope
    (`ApiEnvelope`, mirroring `pkg.APIResponse`):
    `{"success": bool, "data": <payload, omitted on error>, "error": <message, omitted
    on success>, "code": <machine-readable error code, omitted on success and on some
    legacy message-only error paths>}`. Success responses set `success: true` and
    populate `data` with an operation-specific payload; error responses set
    `success: false`, `error` (safe to display for 4xx, generic "internal server error"
    text for 5xx — see `pkg.Error`), and `code` (one of the `ErrorCode` enum values).

    ## Permission model
    Server-scoped authorization (`models.Permission` in `server/models/role.go`) is a
    20-bit bitfield stored per `Role`; a member's effective permission mask is the
    bitwise OR of every role they hold. Endpoints gated on one specific bit at the
    routing-middleware layer (`authServerPerm` in `init_routes.go`) carry the
    `x-required-permission` extension (bit name + decimal value). `PermAdmin` (512)
    short-circuits every check in `Permission.Has` — a role holding `PermAdmin` passes
...

...

...

    ## Auth
...

...
servers:
...
security:
  - bearerAuth: []
  - botAuth: []
paths:
...
  /api/servers:
    get:
      operationId: serversList
      summary: List servers the caller belongs to.
      tags: [Servers]
      responses:
        '200': { $ref: '#/components/responses/StubOK' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
    post:
      operationId: serversCreate
      summary: Create a new server.
      tags: [Servers]
      requestBody: { $ref: '#/components/requestBodies/StubJSON' }
      responses:
        '201': { $ref: '#/components/responses/StubCreated' }
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
...
  /api/admin/servers:
    get:
      operationId: adminListServers
      summary: List all servers on the platform.
      tags: [Admin]
      x-platform-admin: true
      responses:
        '200': { $ref: '#/components/responses/StubOK' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
...
  /api/admin/servers/{serverId}:
    delete:
      operationId: adminDeleteServer
      summary: Hard-delete a server (platform-admin override).
      tags: [Admin]
      x-platform-admin: true
      parameters:
        - $ref: '#/components/parameters/ServerId'
      requestBody: { $ref: '#/components/requestBodies/StubJSON' }
      responses:
        '200': { $ref: '#/components/responses/StubOK' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }

  # ── Platform Admin — Reports ──
...
  /api/servers/{serverId}:
    get:
      operationId: serversGet
      summary: Get a server's details.
      tags: [Servers]
      parameters:
        - $ref: '#/components/parameters/ServerId'
      responses:
        '200': { $ref: '#/components/responses/StubOK' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    patch:
      operationId: serversUpdate
      summary: Update a server's settings.
      tags: [Servers]
      x-required-permission: { name: PermAdmin, value: 512 }
      parameters:
        - $ref: '#/components/parameters/ServerId'
      requestBody: { $ref: '#/components/requestBodies/StubJSON' }
      responses:
        '200': { $ref: '#/components/responses/StubOK' }
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    delete:
      operationId: serversDelete
      summary: Delete a server (owner action).
      tags: [Servers]
      parameters:
        - $ref: '#/components/parameters/ServerId'
      responses:
        '200': { $ref: '#/components/responses/StubOK' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
...
  /ws:
    get:
      operationId: wsConnect
      summary: Upgrade to the human realtime gateway WebSocket connection.
      description: >-
        This is a WebSocket handshake (HTTP Upgrade), not a plain HTTP request —
        modeled here only so the route is enumerated. Preferred auth: obtain a
        one-shot, ~30s-TTL ticket from POST /api/auth/ws-ticket, then connect
        with `?ticket=<value>`. A legacy `?token=<access_token>` query-string
        path exists but is disabled server-side pending removal. No
        Authorization header is used on this route — the ticket/token is a
        query parameter because browsers cannot set custom headers on a
        WebSocket handshake.
      tags: [WebSocket]
      security: []
      parameters:
        - name: ticket
          in: query
          required: false
          schema: { type: string }
          description: One-shot ticket minted by POST /api/auth/ws-ticket.
      responses:
        '101':
          description: Switching Protocols — WebSocket handshake accepted.
        '401': { $ref: '#/components/responses/UnauthorizedError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Human session access token, minted by POST /api/auth/login or
        POST /api/auth/refresh. Sent as `Authorization: Bearer <jwt>`.
    botAuth:
      type: http
      scheme: bearer
      description: >-
        Bot automation credential minted by POST /api/bots. Uses the identical
        `Authorization: Bearer <token>` header as bearerAuth; AuthMiddleware.Require
        distinguishes the two by the models.BotTokenPrefix `hb_` prefix on the
        token value itself, before falling back to JWT validation. Accepted on
        every route documented with the default bearerAuth + botAuth security
        requirement (both are alternatives — either satisfies it).
    mediaCookie:
      type: apiKey
      in: cookie
      name: hichat_media
      description: >-
        Scope-limited media JWT set as an HttpOnly cookie on login/refresh.
        Authenticates ONLY GET /api/uploads/* (channel/DM attachment bytes),
        for contexts that cannot set a header — a bare `<img>`/`<video>` src.
        AuthMiddleware rejects any non-empty token scope on every other route,
        so a leaked media token cannot be replayed against the rest of the API.
    livekitWebhook:
      type: apiKey
      in: header
      name: Authorize
      description: >-
        Not a per-caller secret in the usual apiKey sense: LiveKit signs each
        webhook delivery with an HMAC-derived, JWT-shaped token in the
        `Authorize` header. The server verifies it against every configured
        LiveKit instance's decrypted API key/secret pair (a multi-key
        webhook.ReceiveWebhookEvent provider), so any known instance's
        credentials validate the signature — there is no client-supplied
        credential to document beyond "this header must verify."
  parameters:
    ServerId:
      name: serverId
      in: path
      required: true
      schema: { type: string }
      description: Server (guild) ID.
...
  schemas:
    ErrorCode:
      type: string
      description: >-
        Machine-readable error identifier derived from the domain sentinel
        error (server/pkg/errors.go) via pkg.errorCode. Empty/omitted on
        success responses and on some legacy message-only error paths
        (pkg.ErrorWithMessage).
      enum:
        - NOT_FOUND
        - UNAUTHORIZED
        - FORBIDDEN
        - ALREADY_EXISTS
        - BAD_REQUEST
        - CONFLICT
        - INVALID_KEY
        - INTERNAL
    ApiEnvelope:
      type: object
      description: Standard response envelope wrapping every API response (pkg.APIResponse).
      properties:
        success:
          type: boolean
          description: true on 2xx responses, false on error responses.
        data:
          description: >-
            Present on success responses; omitted on error responses. Payload
            shape is operation-specific — see the allOf composition on each
            2xx response.
        error:
          type: string
          description: Human-readable message. Present only on error responses (safe to display for 4xx; generic text for 5xx).
        code:
          $ref: '#/components/schemas/ErrorCode'
      required: [success]
    ErrorResponse:
      description: Envelope shape on error responses — success is always false; data is omitted.
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            success:
              type: boolean
              enum: [false]
          required: [success, error]
    StubData:
      description: 'TODO: detailed schema — to be specified in the next authoring phase.'
      type: object
      additionalProperties: true

    # ── Shared primitives (Auth / Users / Messages / Reactions / Pins / Channels / Categories) ──
...
  responses:
    StubOK:
      description: Success (200). Payload is a stub in this phase — see StubData.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiEnvelope'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/StubData'
    StubCreated:
      description: Resource created (201). Payload is a stub in this phase — see StubData.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiEnvelope'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/StubData'
...
    BadRequestError:
      description: Malformed request — validation failure, missing required field, or invalid query/path parameter.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UnauthorizedError:
      description: Missing, malformed, expired, or revoked credential (session JWT or bot token).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ForbiddenError:
      description: >-
        Authenticated but not authorized — not a server member, missing the
        required permission bit, or (on platform routes) not a platform admin.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFoundError:
      description: Target resource does not exist, or is not visible to the caller.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  requestBodies:
    StubJSON:
      description: 'JSON request payload (stub — fields to be specified in the next authoring phase).'
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StubData' }
...